Add auicgp instr, add commented-out auipcc instr
CHERIoT supports the `auicgp` instruction suggested but not defined
in the "Capability Hardware Enhanced RISC Instructions:
CHERI Instruction-Set Architecture (Version 9)" technical report.
Add this instruction so it can be tested in future.
Also add in commented-out form the `auipcc` instruction that shares
its encoding with the `auipc` instruction, for documentation purposes.