Improve interrupt testing
Improve interrupt testing by switching from
Machine External Interrupt to Machine Software Interrupt
and adding a randomised version of the test.
Using MSI allows the interrupt to be cleared using the
memory-mapped CLINT interface that is not normally
enabled in sail-riscv when building for RVFI_DII.
The randomised version of the test simply inserts random
instructions between blocks of `noShrink`-ed
interrupt-related instructions.